feat: The custom hook useFetchGifs was created
